requirements.txt
langchain 
langchain-community
langchain-core
langchain-huggingface
pypdf
tiktoken
aiofiles
fastapi
uvicorn
jinja2
python-multipart
PyPDF2
faiss-cpu
python-dotenv
langchain-groq

